Refactoring: Auth.php is deleted, code added to locallib.php
